+++ src/main/kotlin/me/ethius/shared/events/def/WindowResizedEvent.kt
@@ -1,6 +1,6 @@
 package me.ethius.shared.events.def
 
+import me.ethius.shared.double
 import me.ethius.shared.events.Event
-import me.ethius.shared.float
 
-data class WindowResizedEvent(val width:float, val height:float):Event()
+data class WindowResizedEvent(val width:double, val height:double):Event()
